Add Rinku for auto linking
Worldwide office comments make use of the `auto_link` method. This used to be
part of Rails but was deprecated.

This adds the Rinku gem, which is a drop in replacement for `auto_link`. 

This matches the implementation of Organisations in the Collections
application. We may be able to refine this further in future by autolinking the
comments before publishing in Whitehall.